Filter transposable elements from rmsk reference based on given their gene, family, and class ID.

For filtering out the select transposable elements from gene, family, and class to down-stream analysis, we provide this function to help quickly list all options out. You can run this function first setting show to TRUE to see every possibility. Then you provide a character vector including the name of family/class you want to select. The function will then return a subset of the rmsk dataframe.

Usage

rmskFilter(rmskInfo, gene = NULL, family = NULL, class = NULL, show = FALSE)

Arguments

rmskInfo

rmsk dataframe, with three columns, gene_id, family_id, and class_id

gene

a vector including the gene names you want

family

a vector including the family names you want

class

a vector including the class names you want

show

if you want to first see all possible family/class, set this as TRUE

Value

a subset of rmsk dataframe including three columns, gene, family and class

Examples

data(hg19rmsk_info)
df <- hg19rmsk_info

#list all possibilities
rmskFilter(df, show=TRUE)
#> [1] "Family ID"
#> [1] "========================="
#>  [1] "Alu"           "hAT-Charlie"   "ERV1"          "MIR"          
#>  [5] "L2"            "ERVL"          "TcMar-Tigger"  "ERVL-MaLR"    
#>  [9] "L1"            "Gypsy"         "CR1"           "TcMar-Mariner"
#> [13] "RNA"           "hAT-Tip100"    "DNA"           "Deu"          
#> [17] "ERVK"          "Helitron"      "TcMar-Tc2"     "hAT-Blackjack"
#> [21] "TcMar"         "SVA_D"         "SVA_C"         "MamRep605"    
#> [25] "hAT"           "RTE-BovB"      "ERV"           "RTE"          
#> [29] "LTR"           "PiggyBac"      "MuDR"          "UCON9"        
#> [33] "Dong-R4"       "telo"          "SINE"          "centr"        
#> [37] "UCON28a"       "UCON2"         "UCON22"        "UCON5"        
#> [41] "MamRep564"     "UCON31"        "SVA_F"         "SVA_B"        
#> [45] "SVA_E"         "SVA_A"         "UCON26"        "Satellite"    
#> [49] "acro"          "UCON18"        "UCON27"        "Penelope"     
#> [53] "UCON4"         "UCON24"        "UCON16"        "UCON11"       
#> [57] "Merlin"        "UCON10"        "MER130"        "Eulor4"       
#> [61] "UCON17"        "UCON6"         "UCON20"        "UCON28c"      
#> [65] "UCON8"         "UCON1"         "UCON19"        "UCON28b"      
#> [69] "UCON12A"       "UCON25"        "UCON15"        "UCON12"       
#> [1] "-------------------------"
#> [1] "Class ID"
#> [1] "========================="
#>  [1] "SINE"      "DNA"       "LTR"       "LINE"      "RNA"       "DNA?"     
#>  [7] "RC"        "Other"     "Unknown"   "Satellite" "SINE?"     "LTR?"     
#> [13] "LINE?"     "Unknown?" 
#>             gene_id     family_id  class_id
#> 1            AluSx3           Alu      SINE
#> 2             MER5A   hAT-Charlie       DNA
#> 3            MER41B          ERV1       LTR
#> 4               MIR           MIR      SINE
#> 5              AluY           Alu      SINE
#> 6              MIRb           MIR      SINE
#> 7               L2a            L2      LINE
#> 8            MER77B          ERVL       LTR
#> 9             LTR44          ERV1       LTR
#> 10            AluJo           Alu      SINE
#> 11         Tigger2a  TcMar-Tigger       DNA
#> 12            MLT1C     ERVL-MaLR       LTR
#> 13            MER1A   hAT-Charlie       DNA
#> 14            L1MEg            L1      LINE
#> 15            AluJr           Alu      SINE
#> 16           AluSc5           Alu      SINE
#> 17            MER5B   hAT-Charlie       DNA
#> 18              L2c            L2      LINE
#> 19            AluSz           Alu      SINE
#> 20           AluSx1           Alu      SINE
#> 21             MIRc           MIR      SINE
#> 22             L1M5            L1      LINE
#> 23            AluSc           Alu      SINE
#> 24             MIR3           MIR      SINE
#> 25          MER106B   hAT-Charlie       DNA
#> 26            L1PB3            L1      LINE
#> 27            AluSx           Alu      SINE
#> 28            L1MA7            L1      LINE
#> 29              L2b            L2      LINE
#> 30            LTR78          ERV1       LTR
#> 31            AluJb           Alu      SINE
#> 32            L1PA5            L1      LINE
#> 33           L1PA11            L1      LINE
#> 34            L1PA4            L1      LINE
#> 35            L1MDa            L1      LINE
#> 36            L1MA4            L1      LINE
#> 37           LTR88b         Gypsy       LTR
#> 38            L1PA8            L1      LINE
#> 39               L3           CR1      LINE
#> 40           L1ME2z            L1      LINE
#> 41             FRAM           Alu      SINE
#> 42           L1PA14            L1      LINE
#> 43            AluYc           Alu      SINE
#> 44           MLT1G3     ERVL-MaLR       LTR
#> 45            MLT2E          ERVL       LTR
#> 46            AluSg           Alu      SINE
#> 47        MER41-int          ERV1       LTR
#> 48            MER20   hAT-Charlie       DNA
#> 49            L1MEe            L1      LINE
#> 50            MLT2F          ERVL       LTR
#> 51           LTR67B          ERVL       LTR
#> 52            THE1B     ERVL-MaLR       LTR
#> 53      MamGypLTR2b         Gypsy       LTR
#> 54            L1PA3            L1      LINE
#> 55            MADE1 TcMar-Mariner       DNA
#> 56       Charlie26a   hAT-Charlie       DNA
#> 57           L1ME4a            L1      LINE
#> 58            MLT1B     ERVL-MaLR       LTR
#> 59               L2            L2      LINE
#> 60            L1MD3            L1      LINE
#> 61           L1PA16            L1      LINE
#> 62              7SK           RNA       RNA
#> 63           MER45B    hAT-Tip100       DNA
#> 64             L1MC            L1      LINE
#> 65           AluSg7           Alu      SINE
#> 66        MLT1F-int     ERVL-MaLR       LTR
#> 67            L1MC1            L1      LINE
#> 68            MLT1F     ERVL-MaLR       LTR
#> 69       MamGypLTR3         Gypsy       LTR
#> 70          Eulor5A           DNA      DNA?
#> 71           AluSq2           Alu      SINE
#> 72         AmnSINE2           Deu      SINE
#> 73           AluSz6           Alu      SINE
#> 74           LTR13A          ERVK       LTR
#> 75            L1MB8            L1      LINE
#> 76            MLT1J     ERVL-MaLR       LTR
#> 77          LTR16B1          ERVL       LTR
#> 78          LTR16B2          ERVL       LTR
#> 79           LTR16B          ERVL       LTR
#> 80             HAL1            L1      LINE
#> 81           MER45A    hAT-Tip100       DNA
#> 82             L1M1            L1      LINE
#> 83  Helitron2Na_Mam      Helitron        RC
#> 84            MARNA TcMar-Mariner       DNA
#> 85           MER91A    hAT-Tip100       DNA
#> 86           AluSc8           Alu      SINE
#> 87          Kanga1d     TcMar-Tc2       DNA
#> 88            L1PB4            L1      LINE
#> 89           FLAM_C           Alu      SINE
#> 90           MER58C   hAT-Charlie       DNA
#> 91            MER30   hAT-Charlie       DNA
#> 92  ERV3-16A3_I-int          ERVL       LTR
#> 93            L1MC3            L1      LINE
#> 94          LTR16E1          ERVL       LTR
#> 95            L1MEa            L1      LINE
#> 96            THE1D     ERVL-MaLR       LTR
#> 97            L1MB2            L1      LINE
#> 98             L1M4            L1      LINE
#> 99            L1MB5            L1      LINE
#> 100           L1MB3            L1      LINE
#> 101           L1MB4            L1      LINE
#> 102            LTR1          ERV1       LTR
#> 103           AluSp           Alu      SINE
#> 104          MER63B hAT-Blackjack       DNA
#> 105           L1MA8            L1      LINE
#> 106          L1PA17            L1      LINE
#> 107            L1M6            L1      LINE
#> 108          AluSg4           Alu      SINE
#> 109       Charlie7a   hAT-Charlie       DNA
#> 110           MER1B   hAT-Charlie       DNA
#> 111            MSTA     ERVL-MaLR       LTR
#> 112           LTR87          ERVL       LTR
#> 113          MER135           DNA       DNA
#> 114          MER58B   hAT-Charlie       DNA
#> 115           L1MC4            L1      LINE
#> 116           MER6B  TcMar-Tigger       DNA
#> 117         Tigger2  TcMar-Tigger       DNA
#> 118          L1PA15            L1      LINE
#> 119         MER103C   hAT-Charlie       DNA
#> 120           L1MCa            L1      LINE
#> 121          MLT2B3          ERVL       LTR
#> 122           LTR39          ERV1       LTR
#> 123         Tigger1  TcMar-Tigger       DNA
#> 124          MER112   hAT-Charlie       DNA
#> 125      MamRep1161         TcMar       DNA
#> 126           MER4C          ERV1       LTR
#> 127          MER11A          ERVK       LTR
#> 128            LTR8          ERV1       LTR
#> 129       Charlie1b   hAT-Charlie       DNA
#> 130           LTR32          ERVL       LTR
#> 131          MER65D          ERV1       LTR
#> 132           MLT1D     ERVL-MaLR       LTR
#> 133          MER117   hAT-Charlie       DNA
#> 134           MLT1A     ERVL-MaLR       LTR
#> 135          MER51E          ERV1       LTR
#> 136          AluJr4           Alu      SINE
#> 137          MLT2C1          ERVL       LTR
#> 138           L1ME1            L1      LINE
#> 139            L1M7            L1      LINE
#> 140       MLT1A-int     ERVL-MaLR       LTR
#> 141           LTR5B          ERVK       LTR
#> 142           L1PA7            L1      LINE
#> 143          L1PA13            L1      LINE
#> 144          MLT2B4          ERVL       LTR
#> 145          MLT1F2     ERVL-MaLR       LTR
#> 146          MER47B  TcMar-Tigger       DNA
#> 147           L1MA1            L1      LINE
#> 148           MER84          ERV1       LTR
#> 149          MLT1A0     ERVL-MaLR       LTR
#> 150           LTR33          ERVL       LTR
#> 151          LTR78B          ERV1       LTR
#> 152       Charlie2b   hAT-Charlie       DNA
#> 153         Arthur1    hAT-Tip100       DNA
#> 154       Charlie2a   hAT-Charlie       DNA
#> 155           SVA_D         SVA_D     Other
#> 156          MER5A1   hAT-Charlie       DNA
#> 157           L1PB1            L1      LINE
#> 158         AluYk11           Alu      SINE
#> 159         LTR5_Hs          ERVK       LTR
#> 160          LTR81B         Gypsy       LTR
#> 161           LTR79          ERVL       LTR
#> 162          MER9a1          ERVK       LTR
#> 163           MLT1H     ERVL-MaLR       LTR
#> 164           MER33   hAT-Charlie       DNA
#> 165        Charlie7   hAT-Charlie       DNA
#> 166           SVA_C         SVA_C     Other
#> 167          MER58A   hAT-Charlie       DNA
#> 168          MER63A hAT-Blackjack       DNA
#> 169           MADE2 TcMar-Mariner       DNA
#> 170          MER44B  TcMar-Tigger       DNA
#> 171         Tigger7  TcMar-Tigger       DNA
#> 172      Charlie17a   hAT-Charlie       DNA
#> 173           L1M4b            L1      LINE
#> 174          AluSx4           Alu      SINE
#> 175      Charlie21a   hAT-Charlie       DNA
#> 176      Charlie15a   hAT-Charlie       DNA
#> 177       MamRep605     MamRep605   Unknown
#> 178          L1ME3C            L1      LINE
#> 179           LTR83          ERVL       LTR
#> 180       Charlie4z   hAT-Charlie       DNA
#> 181          MLT1J2     ERVL-MaLR       LTR
#> 182          FLAM_A           Alu      SINE
#> 183       Tigger12A  TcMar-Tigger       DNA
#> 184           L1ME5            L1      LINE
#> 185           L1MD2            L1      LINE
#> 186           L1MB7            L1      LINE
#> 187           L1MEc            L1      LINE
#> 188          L1ME3F            L1      LINE
#> 189       Tigger15a  TcMar-Tigger       DNA
#> 190       MamRep137         TcMar       DNA
#> 191          MER110          ERV1       LTR
#> 192          L1ME3D            L1      LINE
#> 193           LTR27          ERV1       LTR
#> 194          LTR48B          ERV1       LTR
#> 195         MER34A1          ERV1       LTR
#> 196       MER31-int          ERV1       LTR
#> 197      MER34B-int          ERV1       LTR
#> 198           LTR45          ERV1       LTR
#> 199          LTR14B          ERVK       LTR
#> 200         MLT1E1A     ERVL-MaLR       LTR
#> 201          MER92C          ERV1       LTR
#> 202          LTR37A          ERV1       LTR
#> 203           THE1C     ERVL-MaLR       LTR
#> 204           AluSq           Alu      SINE
#> 205         AluSq10           Alu      SINE
#> 206          MER34A          ERV1       LTR
#> 207           LTR6B          ERV1       LTR
#> 208          AluYa5           Alu      SINE
#> 209       MLT1I-int     ERVL-MaLR       LTR
#> 210           LTR23          ERV1       LTR
#> 211           MLT1I     ERVL-MaLR       LTR
#> 212     ERVL-B4-int          ERVL       LTR
#> 213           MER53           hAT       DNA
#> 214           LTR49          ERV1       LTR
#> 215           THE1A     ERVL-MaLR       LTR
#> 216       MamRep434  TcMar-Tigger       DNA
#> 217          L1MC4a            L1      LINE
#> 218           L1MC5            L1      LINE
#> 219           MLT1L     ERVL-MaLR       LTR
#> 220          L1MA4A            L1      LINE
#> 221         MER113A   hAT-Charlie       DNA
#> 222            MER3   hAT-Charlie       DNA
#> 223       Tigger12c  TcMar-Tigger       DNA
#> 224           LTR1D          ERV1       LTR
#> 225            MER8  TcMar-Tigger       DNA
#> 226          MLT1N2     ERVL-MaLR       LTR
#> 227            MER2  TcMar-Tigger       DNA
#> 228          MLT1F1     ERVL-MaLR       LTR
#> 229           L1PB2            L1      LINE
#> 230           L1MA3            L1      LINE
#> 231            L1M2            L1      LINE
#> 232           MER81 hAT-Blackjack       DNA
#> 233      Charlie18a   hAT-Charlie       DNA
#> 234          AluYf4           Alu      SINE
#> 235        Charlie8   hAT-Charlie       DNA
#> 236      Charlie22a   hAT-Charlie       DNA
#> 237           L1MEf            L1      LINE
#> 238          MLT1J1     ERVL-MaLR       LTR
#> 239          MER90a          ERV1       LTR
#> 240          L1PA10            L1      LINE
#> 241           L1ME3            L1      LINE
#> 242           LTR51          ERV1       LTR
#> 243            MSTB     ERVL-MaLR       LTR
#> 244           MER39          ERV1       LTR
#> 245        Tigger3a  TcMar-Tigger       DNA
#> 246          MER104     TcMar-Tc2       DNA
#> 247          LTR40a          ERVL       LTR
#> 248       Tigger16b  TcMar-Tigger       DNA
#> 249             FAM           Alu      SINE
#> 250            L1P4            L1      LINE
#> 251          AluSq4           Alu      SINE
#> 252          MER66B          ERV1       LTR
#> 253          MER31B          ERV1       LTR
#> 254       THE1B-int     ERVL-MaLR       LTR
#> 255            L1MD            L1      LINE
#> 256            L1P5            L1      LINE
#> 257        Tigger3d  TcMar-Tigger       DNA
#> 258          AluYd8           Alu      SINE
#> 259            L1M3            L1      LINE
#> 260          AluYg6           Alu      SINE
#> 261             L3b           CR1      LINE
#> 262           MLT1K     ERVL-MaLR       LTR
#> 263         X3_LINE      RTE-BovB      LINE
#> 264           MER5C   hAT-Charlie       DNA
#> 265        Tigger4b  TcMar-Tigger       DNA
#> 266         LTR16A1          ERVL       LTR
#> 267          LTR24C          ERV1       LTR
#> 268        X7C_LINE           CR1      LINE
#> 269         Tigger5  TcMar-Tigger       DNA
#> 270          LTR10C          ERV1       LTR
#> 271      HAL1-3A_ME            L1      LINE
#> 272          MER44A  TcMar-Tigger       DNA
#> 273           MER95           ERV       LTR
#> 274          MER34C          ERV1       LTR
#> 275          MER21B          ERVL       LTR
#> 276         MER102c   hAT-Charlie       DNA
#> 277         L1PREC2            L1      LINE
#> 278          LTR33A          ERVL       LTR
#> 279    HUERS-P1-int          ERV1       LTR
#> 280      ERVL-E-int          ERVL       LTR
#> 281          MLT2A2          ERVL       LTR
#> 282          MER47A  TcMar-Tigger       DNA
#> 283          MLT2B2          ERVL       LTR
#> 284          MLT2B1          ERVL       LTR
#> 285          MER83B          ERV1       LTR
#> 286          MER52A          ERV1       LTR
#> 287          L1ME3E            L1      LINE
#> 288          MER21C          ERVL       LTR
#> 289          MER51A          ERV1       LTR
#> 290        MSTB-int     ERVL-MaLR       LTR
#> 291           MLT1M     ERVL-MaLR       LTR
#> 292              L4           RTE      LINE
#> 293          LTR16C          ERVL       LTR
#> 294           MER50          ERV1       LTR
#> 295       MER50-int          ERV1       LTR
#> 296          MER4D1          ERV1       LTR
#> 297          LTR88c         Gypsy       LTR
#> 298            MSTD     ERVL-MaLR       LTR
#> 299    Tigger2b_Pri  TcMar-Tigger       DNA
#> 300         Plat_L3           CR1      LINE
#> 301            LTR9          ERV1       LTR
#> 302           MSTB1     ERVL-MaLR       LTR
#> 303       Charlie24   hAT-Charlie       DNA
#> 304        MamRep38           hAT       DNA
#> 305          L1MA5A            L1      LINE
#> 306        Arthur1A    hAT-Tip100       DNA
#> 307          MER124           DNA      DNA?
#> 308          MER113   hAT-Charlie       DNA
#> 309          MER63C hAT-Blackjack       DNA
#> 310       Charlie4a   hAT-Charlie       DNA
#> 311        X6B_LINE           CR1      LINE
#> 312             L1M            L1      LINE
#> 313           L1ME2            L1      LINE
#> 314           L1PA6            L1      LINE
#> 315         LTR86B2          ERVL       LTR
#> 316        Tigger3b  TcMar-Tigger       DNA
#> 317           L1MEd            L1      LINE
#> 318           ALINE           RTE      LINE
#> 319      MamRep1527           LTR       LTR
#> 320           MER68          ERVL       LTR
#> 321          L1MA10            L1      LINE
#> 322            MSTC     ERVL-MaLR       LTR
#> 323         ORSL-2a    hAT-Tip100       DNA
#> 324          MER115    hAT-Tip100       DNA
#> 325           MER75      PiggyBac       DNA
#> 326      PRIMA4-int          ERV1       LTR
#> 327      PRIMAX-int          ERV1       LTR
#> 328       THE1D-int     ERVL-MaLR       LTR
#> 329          MLT2A1          ERVL       LTR
#> 330       HERVL-int          ERVL       LTR
#> 331         Kanga1c     TcMar-Tc2       DNA
#> 332           L1MC2            L1      LINE
#> 333          LTR85b         Gypsy       LTR
#> 334      MLT1A0-int     ERVL-MaLR       LTR
#> 335            L1P3            L1      LINE
#> 336           L1M3d            L1      LINE
#> 337       MLT1J-int     ERVL-MaLR       LTR
#> 338           L1MA2            L1      LINE
#> 339          MER96B           hAT       DNA
#> 340      Charlie13b   hAT-Charlie       DNA
#> 341            L1PB            L1      LINE
#> 342           L1PBa            L1      LINE
#> 343          MLT1G1     ERVL-MaLR       LTR
#> 344           LTR12          ERV1       LTR
#> 345           MLT2D          ERVL       LTR
#> 346          L1ME3A            L1      LINE
#> 347          LTR81C         Gypsy       LTR
#> 348           L1MB1            L1      LINE
#> 349          MLT1H2     ERVL-MaLR       LTR
#> 350           L1MA5            L1      LINE
#> 351          LTR54B          ERV1       LTR
#> 352           MER72          ERV1       LTR
#> 353         MER102b   hAT-Charlie       DNA
#> 354          L1ME3B            L1      LINE
#> 355         LTR33A_          ERVL       LTR
#> 356     MamGypLTR1c         Gypsy       LTR
#> 357         LTR16A2          ERVL       LTR
#> 358       Tigger16a  TcMar-Tigger       DNA
#> 359         LTR16E2          ERVL       LTR
#> 360           LTR8A          ERV1       LTR
#> 361           L1MCc            L1      LINE
#> 362           L1M4c            L1      LINE
#> 363          MER105           DNA       DNA
#> 364 Helitron3Na_Mam      Helitron        RC
#> 365        X7A_LINE           CR1      LINE
#> 366        X7B_LINE           CR1      LINE
#> 367        Charlie5   hAT-Charlie       DNA
#> 368       Charlie10   hAT-Charlie       DNA
#> 369          MER107           DNA       DNA
#> 370         Kanga1b     TcMar-Tc2       DNA
#> 371        MSTA-int     ERVL-MaLR       LTR
#> 372          MER97a    hAT-Tip100       DNA
#> 373           MER85      PiggyBac       DNA
#> 374      Charlie16a   hAT-Charlie       DNA
#> 375          MER119   hAT-Charlie       DNA
#> 376           MSTB2     ERVL-MaLR       LTR
#> 377          LTR43B          ERV1       LTR
#> 378           L1PA2            L1      LINE
#> 379       MamRep488           hAT       DNA
#> 380          LTR16D          ERVL       LTR
#> 381          MLT1E2     ERVL-MaLR       LTR
#> 382           MER94 hAT-Blackjack       DNA
#> 383           HAL1b            L1      LINE
#> 384          MER46C  TcMar-Tigger       DNA
#> 385          MLT1H1     ERVL-MaLR       LTR
#> 386          LTR16A          ERVL       LTR
#> 387          LTR47A          ERVL       LTR
#> 388       LTR49-int          ERV1       LTR
#> 389       Tigger11a  TcMar-Tigger       DNA
#> 390          MLT1A1     ERVL-MaLR       LTR
#> 391        Kanga11a     TcMar-Tc2       DNA
#> 392      Charlie10a   hAT-Charlie       DNA
#> 393          LTR26E          ERV1       LTR
#> 394   ERV3-16A3_LTR          ERVL       LTR
#> 395      Charlie23a   hAT-Charlie       DNA
#> 396          MER51B          ERV1       LTR
#> 397          MER61B          ERV1       LTR
#> 398         MER110A          ERV1       LTR
#> 399     MamGypLTR1b         Gypsy       LTR
#> 400       THE1A-int     ERVL-MaLR       LTR
#> 401      MamRep1879           hAT       DNA
#> 402          MER4D0          ERV1       LTR
#> 403       MER4B-int          ERV1       LTR
#> 404         MLT-int     ERVL-MaLR       LTR
#> 405       Ricksha_c          MuDR       DNA
#> 406          L1PA12            L1      LINE
#> 407          MER21A          ERVL       LTR
#> 408       MER21-int          ERVL       LTR
#> 409           MER4A          ERV1       LTR
#> 410       BLACKJACK hAT-Blackjack       DNA
#> 411          LTR33B          ERVL       LTR
#> 412           LTR41          ERVL       LTR
#> 413      MER110-int          ERV1       LTR
#> 414           MER4D          ERV1       LTR
#> 415          MER50B          ERV1       LTR
#> 416           LTR36          ERV1       LTR
#> 417         MER133A           DNA      DNA?
#> 418         MER57C2          ERV1       LTR
#> 419      MER57A-int          ERV1       LTR
#> 420           LTR48          ERV1       LTR
#> 421          LTR12F          ERV1       LTR
#> 422       HERV9-int          ERV1       LTR
#> 423  HERVIP10FH-int          ERV1       LTR
#> 424          LTR12B          ERV1       LTR
#> 425        ERVL-int          ERVL       LTR
#> 426          LTR37B          ERV1       LTR
#> 427          MLT2B5          ERVL       LTR
#> 428          LTR85c         Gypsy       LTR
#> 429           LTR50          ERVL       LTR
#> 430           MER77          ERVL       LTR
#> 431          MER11D          ERVK       LTR
#> 432          LTR24B          ERV1       LTR
#> 433          AluYk4           Alu      SINE
#> 434          LTR33C          ERVL       LTR
#> 435         Tigger4  TcMar-Tigger       DNA
#> 436           LTR53          ERVL       LTR
#> 437          MER39B          ERV1       LTR
#> 438           UCON9         UCON9   Unknown
#> 439            LTR7          ERV1       LTR
#> 440       HERVH-int          ERV1       LTR
#> 441           L1MA9            L1      LINE
#> 442           L1MD1            L1      LINE
#> 443        Arthur1B    hAT-Tip100       DNA
#> 444     MamGypLTR1a         Gypsy       LTR
#> 445          MER31A          ERV1       LTR
#> 446          MER67B          ERV1       LTR
#> 447          MER67A          ERV1       LTR
#> 448          LTR10A          ERV1       LTR
#> 449          MER41C          ERV1       LTR
#> 450          MER94B hAT-Blackjack       DNA
#> 451           LTR54          ERV1       LTR
#> 452          MLT1E3     ERVL-MaLR       LTR
#> 453           MLT1G     ERVL-MaLR       LTR
#> 454      Charlie19a   hAT-Charlie       DNA
#> 455      MLT1J2-int     ERVL-MaLR       LTR
#> 456           MER82  TcMar-Tigger       DNA
#> 457         Eulor2B           DNA      DNA?
#> 458          L1PA8A            L1      LINE
#> 459           LTR6A          ERV1       LTR
#> 460        Tigger4a  TcMar-Tigger       DNA
#> 461           LTR65          ERV1       LTR
#> 462        Tigger9a  TcMar-Tigger       DNA
#> 463       THE1C-int     ERVL-MaLR       LTR
#> 464        Charlie9   hAT-Charlie       DNA
#> 465      Charlie20a   hAT-Charlie       DNA
#> 466          MER20B   hAT-Charlie       DNA
#> 467          Mam_R4       Dong-R4      LINE
#> 468      MLT1G1-int     ERVL-MaLR       LTR
#> 469          L1MEg1            L1      LINE
#> 470        Charlie3   hAT-Charlie       DNA
#> 471      MLT1F2-int     ERVL-MaLR       LTR
#> 472          MER50C          ERV1       LTR
#> 473           L1PBb            L1      LINE
#> 474        Tigger10  TcMar-Tigger       DNA
#> 475         Ricksha          MuDR       DNA
#> 476         LTR10B1          ERV1       LTR
#> 477       HERVI-int          ERV1       LTR
#> 478       LTR23-int          ERV1       LTR
#> 479     HERV35I-int          ERV1       LTR
#> 480           LTR29          ERV1       LTR
#> 481          MER4A1          ERV1       LTR
#> 482        MER4-int          ERV1       LTR
#> 483       Ricksha_0          MuDR       DNA
#> 484           LTR26          ERV1       LTR
#> 485           L1MEb            L1      LINE
#> 486         MER34C2          ERV1       LTR
#> 487        Tigger6a  TcMar-Tigger       DNA
#> 488      MLT1A1-int     ERVL-MaLR       LTR
#> 489          LTR80B          ERVL       LTR
#> 490          MER67D          ERV1       LTR
#> 491          MER61C          ERV1       LTR
#> 492          L1MEg2            L1      LINE
#> 493          MER54A          ERVL       LTR
#> 494          AluYb8           Alu      SINE
#> 495            L1HS            L1      LINE
#> 496          LTR12C          ERV1       LTR
#> 497            L1P1            L1      LINE
#> 498           LTR43          ERV1       LTR
#> 499           LTR89          ERVL       LTR
#> 500           L1M3e            L1      LINE
#> 501           MLT1E     ERVL-MaLR       LTR
#> 502       Charlie1a   hAT-Charlie       DNA
#> 503          MER67C          ERV1       LTR
#> 504    HUERS-P3-int          ERV1       LTR
#> 505          MLT1E1     ERVL-MaLR       LTR
#> 506           L1M3c            L1      LINE
#> 507       L1PA15-16            L1      LINE
#> 508          L1M2a1            L1      LINE
#> 509          MER61E          ERV1       LTR
#> 510           LTR25          ERV1       LTR
#> 511          MER11B          ERVK       LTR
#> 512       MLT1B-int     ERVL-MaLR       LTR
#> 513        MLT1-int     ERVL-MaLR       LTR
#> 514           L1MA6            L1      LINE
#> 515          MER92B          ERV1       LTR
#> 516        MSTD-int     ERVL-MaLR       LTR
#> 517        AmnSINE1           Deu      SINE
#> 518       Tigger14a  TcMar-Tigger       DNA
#> 519        Kanga2_a     TcMar-Tc2       DNA
#> 520       MLT1C-int     ERVL-MaLR       LTR
#> 521          LTR12_          ERV1       LTR
#> 522           L1P4a            L1      LINE
#> 523         LTR40A1          ERVL       LTR
#> 524       LTR52-int          ERVL       LTR
#> 525           LTR55           ERV       LTR
#> 526         MER57E1          ERV1       LTR
#> 527          AluYc3           Alu      SINE
#> 528           MER34          ERV1       LTR
#> 529        Tigger3c  TcMar-Tigger       DNA
#> 530           LTR60          ERV1       LTR
#> 531            TAR1          telo Satellite
#> 532          MER41A          ERV1       LTR
#> 533          MER45C    hAT-Tip100       DNA
#> 534            LTR3          ERVK       LTR
#> 535      HERVK3-int          ERVK       LTR
#> 536         AluYk12           Alu      SINE
#> 537         LTR16D2          ERVL       LTR
#> 538        X7D_LINE           CR1      LINE
#> 539            MER6  TcMar-Tigger       DNA
#> 540     LFSINE_Vert          SINE      SINE
#> 541         Eulor11           DNA       DNA
#> 542      MLT1H1-int     ERVL-MaLR       LTR
#> 543             L1P            L1      LINE
#> 544           MER96           hAT       DNA
#> 545          MER57D          ERV1       LTR
#> 546          MER63D hAT-Blackjack       DNA
#> 547 Helitron1Na_Mam      Helitron        RC
#> 548       MSTB1-int     ERVL-MaLR       LTR
#> 549         MER106A   hAT-Charlie       DNA
#> 550           LTR77          ERV1       LTR
#> 551      HERV16-int          ERVL       LTR
#> 552      MamGyp-int         Gypsy       LTR
#> 553     FordPrefect    hAT-Tip100       DNA
#> 554         Kanga1a     TcMar-Tc2       DNA
#> 555          Kanga1     TcMar-Tc2       DNA
#> 556      MLT1H2-int     ERVL-MaLR       LTR
#> 557      MamRep1894           hAT       DNA
#> 558          MER70B          ERVL       LTR
#> 559       MER61-int          ERV1       LTR
#> 560       LTR25-int          ERV1       LTR
#> 561       Tigger13a  TcMar-Tigger       DNA
#> 562          LTR80A          ERVL       LTR
#> 563           LTR2B          ERV1       LTR
#> 564   Harlequin-int          ERV1       LTR
#> 565     HERVE_a-int          ERV1       LTR
#> 566       ALR/Alpha         centr Satellite
#> 567         LTR86B1          ERVL       LTR
#> 568          LTR84b          ERVL       LTR
#> 569           LTR7B          ERV1       LTR
#> 570           MER49          ERV1       LTR
#> 571          MER70A          ERVL       LTR
#> 572       Charlie25   hAT-Charlie       DNA
#> 573          LTR86C          ERVL       LTR
#> 574          MER11C          ERVK       LTR
#> 575           LTR52          ERVL       LTR
#> 576           MER4B          ERV1       LTR
#> 577      MER83B-int          ERV1       LTR
#> 578       MER57-int          ERV1       LTR
#> 579          MER65A          ERV1       LTR
#> 580           LTR13          ERVK       LTR
#> 581          LTR84a          ERVL       LTR
#> 582           MER2B  TcMar-Tigger       DNA
#> 583           LTR7C          ERV1       LTR
#> 584          LTR12D          ERV1       LTR
#> 585          MER52C          ERV1       LTR
#> 586            LTR2          ERV1       LTR
#> 587          LTR19A          ERV1       LTR
#> 588          MER9a3          ERVK       LTR
#> 589          MER47C  TcMar-Tigger       DNA
#> 590         MER57B2          ERV1       LTR
#> 591          Zaphod    hAT-Tip100       DNA
#> 592          MER44C  TcMar-Tigger       DNA
#> 593         Zaphod3    hAT-Tip100       DNA
#> 594          LTR3B_          ERVK       LTR
#> 595         LTR86A1          ERVL       LTR
#> 596          HSMAR1 TcMar-Mariner       DNA
#> 597           LTR66          ERVL       LTR
#> 598           LTR81         Gypsy       LTR
#> 599   HERVIP10F-int          ERV1       LTR
#> 600           MER76          ERVL       LTR
#> 601         MER101B          ERV1       LTR
#> 602      MER101-int          ERV1       LTR
#> 603        THE1-int     ERVL-MaLR       LTR
#> 604            ORSL    hAT-Tip100       DNA
#> 605          LTR71A          ERV1       LTR
#> 606          MER74A          ERVL       LTR
#> 607           LTR68          ERV1       LTR
#> 608          MER72B          ERV1       LTR
#> 609       MLT1D-int     ERVL-MaLR       LTR
#> 610           L1M3f            L1      LINE
#> 611         MER57A1          ERV1       LTR
#> 612       MER51-int          ERV1       LTR
#> 613          MLT2C2          ERVL       LTR
#> 614           LTR46          ERV1       LTR
#> 615          L1PBa1            L1      LINE
#> 616          LTR18B          ERVL       LTR
#> 617     HERVL18-int          ERVL       LTR
#> 618          AluYb9           Alu      SINE
#> 619          MER5C1   hAT-Charlie       DNA
#> 620          MER91C    hAT-Tip100       DNA
#> 621           LTR15          ERV1       LTR
#> 622           LTR1C          ERV1       LTR
#> 623    HUERS-P2-int          ERV1       LTR
#> 624          MER66C          ERV1       LTR
#> 625          UCON14           DNA      DNA?
#> 626       MER89-int          ERV1       LTR
#> 627           L1MCb            L1      LINE
#> 628           LTR76          ERV1       LTR
#> 629           LTR9B          ERV1       LTR
#> 630       MER34-int          ERV1       LTR
#> 631       MLT1H-int     ERVL-MaLR       LTR
#> 632           LOR1b          ERV1       LTR
#> 633          UCON29      PiggyBac       DNA
#> 634          HSMAR2 TcMar-Mariner       DNA
#> 635         MER4A1_          ERV1       LTR
#> 636      MER83A-int          ERV1       LTR
#> 637          LTR38B          ERV1       LTR
#> 638         Eulor6D           DNA      DNA?
#> 639      Charlie13a   hAT-Charlie       DNA
#> 640          MER74B          ERVL       LTR
#> 641        LOR1-int          ERV1       LTR
#> 642          MER61A          ERV1       LTR
#> 643            L1P2            L1      LINE
#> 644        Tigger9b  TcMar-Tigger       DNA
#> 645          MER52D          ERV1       LTR
#> 646          MER97b    hAT-Tip100       DNA
#> 647           MER99           hAT       DNA
#> 648           MER6A  TcMar-Tigger       DNA
#> 649     HERVK11-int          ERVK       LTR
#> 650        MSTC-int     ERVL-MaLR       LTR
#> 651     MamGypLTR1d         Gypsy       LTR
#> 652          MER92A          ERV1       LTR
#> 653         CR1_Mam           CR1      LINE
#> 654        Tigger6b  TcMar-Tigger       DNA
#> 655          LTR81A         Gypsy       LTR
#> 656   HUERS-P3b-int          ERV1       LTR
#> 657          LTR21B          ERV1       LTR
#> 658          MER126           DNA       DNA
#> 659         Tigger3  TcMar-Tigger       DNA
#> 660         MER102a   hAT-Charlie       DNA
#> 661          MER97d    hAT-Tip100       DNA
#> 662          MER131          SINE     SINE?
#> 663           MER73          ERVL       LTR
#> 664     HERVL74-int          ERVL       LTR
#> 665          MER44D  TcMar-Tigger       DNA
#> 666           LTR57          ERVL       LTR
#> 667          LTR10D          ERV1       LTR
#> 668         UCON28a       UCON28a   Unknown
#> 669           UCON2         UCON2   Unknown
#> 670          UCON22        UCON22   Unknown
#> 671           UCON5         UCON5   Unknown
#> 672       MamRep564     MamRep564   Unknown
#> 673        Tigger12  TcMar-Tigger       DNA
#> 674          UCON31        UCON31   Unknown
#> 675          MER68B          ERVL       LTR
#> 676          MER83C          ERV1       LTR
#> 677        Charlie1   hAT-Charlie       DNA
#> 678         LTR16D1          ERVL       LTR
#> 679          LTR85a         Gypsy       LTR
#> 680          LTR45C          ERV1       LTR
#> 681           MER89          ERV1       LTR
#> 682           L1M3b            L1      LINE
#> 683           LTR17          ERV1       LTR
#> 684           SVA_F         SVA_F     Other
#> 685          MER58D   hAT-Charlie       DNA
#> 686        Cheshire   hAT-Charlie       DNA
#> 687      MLT1J1-int     ERVL-MaLR       LTR
#> 688      HAL1-2a_MD            L1      LINE
#> 689           MER83          ERV1       LTR
#> 690           LTR56          ERV1       LTR
#> 691          MER75A      PiggyBac       DNA
#> 692       MER84-int          ERV1       LTR
#> 693           MER4E          ERV1       LTR
#> 694         Zaphod2    hAT-Tip100       DNA
#> 695           SVA_B         SVA_B     Other
#> 696          MER121         TcMar       DNA
#> 697          AluYf5           Alu      SINE
#> 698         Tigger8  TcMar-Tigger       DNA
#> 699        Charlie4   hAT-Charlie       DNA
#> 700         ORSL-2b    hAT-Tip100       DNA
#> 701          PABL_A          ERV1       LTR
#> 702      MLT1N2-int     ERVL-MaLR       LTR
#> 703      PABL_A-int          ERV1       LTR
#> 704          MER91B    hAT-Tip100       DNA
#> 705      HERV17-int          ERV1       LTR
#> 706         MER57B1          ERV1       LTR
#> 707      MamRep4096           hAT       DNA
#> 708   FordPrefect_a    hAT-Tip100       DNA
#> 709           L1M2a            L1      LINE
#> 710           MER88          ERVL       LTR
#> 711       MER65-int          ERV1       LTR
#> 712       LTR57-int          ERVL       LTR
#> 713          LTR82A          ERVL       LTR
#> 714           LTR1B          ERV1       LTR
#> 715          LTR22A          ERVK       LTR
#> 716          MER4E1          ERV1       LTR
#> 717     HERVL66-int          ERVL       LTR
#> 718          LTR40b          ERVL       LTR
#> 719         MER113B   hAT-Charlie       DNA
#> 720          AluYh9           Alu      SINE
#> 721           LOR1a          ERV1       LTR
#> 722      PRIMA4_LTR          ERV1       LTR
#> 723          MER51C          ERV1       LTR
#> 724           UCON7           DNA      DNA?
#> 725           SVA_E         SVA_E     Other
#> 726              L5           RTE      LINE
#> 727           SVA_A         SVA_A     Other
#> 728           LTR5A          ERVK       LTR
#> 729        Arthur1C    hAT-Tip100       DNA
#> 730           LTR2C          ERV1       LTR
#> 731          MER57F          ERV1       LTR
#> 732          MER34B          ERV1       LTR
#> 733      Charlie14a   hAT-Charlie       DNA
#> 734          LTR90A           LTR       LTR
#> 735          LTR90B           LTR       LTR
#> 736          Looper      PiggyBac       DNA
#> 737            MERX  TcMar-Tigger       DNA
#> 738          LTR41B          ERVL       LTR
#> 739          LTR26B          ERV1       LTR
#> 740          AluYa8           Alu      SINE
#> 741           L1M2c            L1      LINE
#> 742          PABL_B          ERV1       LTR
#> 743          LTR88a         Gypsy       LTR
#> 744          MER41D          ERV1       LTR
#> 745          LTR47B          ERVL       LTR
#> 746       MER68-int          ERVL       LTR
#> 747          LTR10F          ERV1       LTR
#> 748          MER30B   hAT-Charlie       DNA
#> 749           LTR38          ERV1       LTR
#> 750       LTR38-int          ERV1       LTR
#> 751           L1M2b            L1      LINE
#> 752           LTR58          ERV1       LTR
#> 753      MLT1G3-int     ERVL-MaLR       LTR
#> 754           MER87          ERV1       LTR
#> 755         Eulor9C           DNA       DNA
#> 756        Charlie6   hAT-Charlie       DNA
#> 757           LTR30          ERV1       LTR
#> 758           LTR24          ERV1       LTR
#> 759     MamGypLTR2c         Gypsy       LTR
#> 760       Ricksha_b          MuDR       DNA
#> 761          MER101          ERV1       LTR
#> 762         X8_LINE           CR1      LINE
#> 763         MER34C_          ERV1       LTR
#> 764          MER54B          ERVL       LTR
#> 765           LTR72          ERV1       LTR
#> 766           L1M3a            L1      LINE
#> 767           LTR28          ERV1       LTR
#> 768          LTR27B          ERV1       LTR
#> 769          UCON26        UCON26   Unknown
#> 770          LTR18A          ERVL       LTR
#> 771          LTR40c          ERVL       LTR
#> 772          MER65C          ERV1       LTR
#> 773         LTR86A2          ERVL       LTR
#> 774          MER97c    hAT-Tip100       DNA
#> 775          LTR13_          ERVK       LTR
#> 776          LTR71B          ERV1       LTR
#> 777          LTR10E          ERV1       LTR
#> 778         X2_LINE           CR1      LINE
#> 779       LTR19-int          ERV1       LTR
#> 780           LTR3B          ERVK       LTR
#> 781           LTR64          ERV1       LTR
#> 782         MER57C1          ERV1       LTR
#> 783           L1P3b            L1      LINE
#> 784    Tigger1a_Art  TcMar-Tigger       DNA
#> 785           LTR61          ERV1       LTR
#> 786          LTR75B          ERVL       LTR
#> 787          MER66A          ERV1       LTR
#> 788          MER41E          ERV1       LTR
#> 789       MER66-int          ERV1       LTR
#> 790       MER52-int          ERV1       LTR
#> 791          LTR22C          ERVK       LTR
#> 792     HERV4_I-int          ERV1       LTR
#> 793     HERVL40-int          ERVL       LTR
#> 794          MER70C          ERVL       LTR
#> 795       HERVK-int          ERVK       LTR
#> 796          MER41G          ERV1       LTR
#> 797          LTR45B          ERV1       LTR
#> 798           LTR7Y          ERV1       LTR
#> 799       HERV3-int          ERV1       LTR
#> 800          GSATII         centr Satellite
#> 801       HERVE-int          ERV1       LTR
#> 802           GSATX         centr Satellite
#> 803          LTR12E          ERV1       LTR
#> 804           LTR75          ERVL       LTR
#> 805          LTR10B          ERV1       LTR
#> 806          LTR14A          ERVK       LTR
#> 807          LTR22B          ERVK       LTR
#> 808           MER6C  TcMar-Tigger       DNA
#> 809        (GAATG)n     Satellite Satellite
#> 810     HERVK14-int          ERVK       LTR
#> 811          MER9a2          ERVK       LTR
#> 812      HERVK9-int          ERVK       LTR
#> 813       LTR39-int          ERV1       LTR
#> 814         MST-int     ERVL-MaLR       LTR
#> 815             CER     Satellite Satellite
#> 816           HSAT5     Satellite Satellite
#> 817           SATR1     Satellite Satellite
#> 818       Ricksha_a          MuDR       DNA
#> 819           ACRO1          acro Satellite
#> 820            LTR4          ERV1       LTR
#> 821          MER51D          ERV1       LTR
#> 822           LTR73          ERV1       LTR
#> 823    HERVP71A-int          ERV1       LTR
#> 824          LTR82B          ERVL       LTR
#> 825           MER90          ERV1       LTR
#> 826      MLT1E3-int     ERVL-MaLR       LTR
#> 827      MLT1F1-int     ERVL-MaLR       LTR
#> 828          HSATII     Satellite Satellite
#> 829      MLT1E2-int     ERVL-MaLR       LTR
#> 830           L1P4d            L1      LINE
#> 831        X6A_LINE           CR1      LINE
#> 832         Eulor9A           DNA       DNA
#> 833           LTR42          ERVL       LTR
#> 834 Helitron1Nb_Mam      Helitron        RC
#> 835      MLT1E1-int     ERVL-MaLR       LTR
#> 836         X1_LINE           CR1      LINE
#> 837       MLT1K-int     ERVL-MaLR       LTR
#> 838          UCON23           DNA      DNA?
#> 839          MER127  TcMar-Tigger       DNA
#> 840           LTR11           ERV       LTR
#> 841          MER61D          ERV1       LTR
#> 842           LTR31          ERV1       LTR
#> 843           MER48          ERV1       LTR
#> 844        Tigger2b  TcMar-Tigger       DNA
#> 845          LTR19B          ERV1       LTR
#> 846       MER70-int          ERVL       LTR
#> 847         MER57E2          ERV1       LTR
#> 848          MER65B          ERV1       LTR
#> 849           LTR14          ERVK       LTR
#> 850    HERVK14C-int          ERVK       LTR
#> 851          MER66D          ERV1       LTR
#> 852     HERVK22-int          ERVK       LTR
#> 853          LTR21A          ERV1       LTR
#> 854    HERVFH21-int          ERV1       LTR
#> 855          MER61F          ERV1       LTR
#> 856         LTR75_1          ERV1       LTR
#> 857       MSTB2-int     ERVL-MaLR       LTR
#> 858           L1MDb            L1      LINE
#> 859       MER76-int          ERVL       LTR
#> 860        DNA1_Mam         TcMar       DNA
#> 861         LTR81AB         Gypsy       LTR
#> 862            LTR5          ERVK       LTR
#> 863           LTR62          ERVL       LTR
#> 864          MER129           LTR      LTR?
#> 865          LTR14C          ERVK       LTR
#> 866          LTR35A          ERV1       LTR
#> 867           LTR34          ERV1       LTR
#> 868          MER45R    hAT-Tip100       DNA
#> 869          LTR35B          ERV1       LTR
#> 870          MER125           DNA       DNA
#> 871          UCON18        UCON18   Unknown
#> 872           L1P4e            L1      LINE
#> 873     PRIMA41-int          ERV1       LTR
#> 874      Charlie10b   hAT-Charlie       DNA
#> 875       PrimLTR79          ERV1       LTR
#> 876         Eulor5B           DNA      DNA?
#> 877         X9_LINE            L1      LINE
#> 878         Eulor12           DNA      DNA?
#> 879          LTR38C          ERV1       LTR
#> 880          LTR19C          ERV1       LTR
#> 881       Charlie11   hAT-Charlie       DNA
#> 882      HERV1_LTRa          ERV1       LTR
#> 883          AluYc5           Alu      SINE
#> 884         Eulor9B           DNA       DNA
#> 885          MER134           DNA      DNA?
#> 886        X5A_LINE           CR1      LINE
#> 887       MLT1L-int     ERVL-MaLR       LTR
#> 888         MER133B           DNA      DNA?
#> 889          UCON27        UCON27   Unknown
#> 890         Eulor2A           DNA      DNA?
#> 891    HERV-Fc2_LTR          ERV1       LTR
#> 892          UCON13      Penelope     LINE?
#> 893           LTR3A          ERVK       LTR
#> 894           UCON4         UCON4   Unknown
#> 895          UCON24        UCON24   Unknown
#> 896          UCON16        UCON16   Unknown
#> 897     HERVS71-int          ERV1       LTR
#> 898     HERVH48-int          ERV1       LTR
#> 899        X5B_LINE           CR1      LINE
#> 900          MER75B      PiggyBac       DNA
#> 901          UCON11        UCON11   Unknown
#> 902          MER74C          ERVL       LTR
#> 903          LTR10G          ERV1       LTR
#> 904          MER87B          ERV1       LTR
#> 905          MER34D          ERV1       LTR
#> 906           LTR69          ERVL       LTR
#> 907      Merlin1_HS        Merlin       DNA
#> 908      PABL_B-int          ERV1       LTR
#> 909          L1M3de            L1      LINE
#> 910           LTR59          ERV1       LTR
#> 911          UCON10        UCON10   Unknown
#> 912          MER130        MER130  Unknown?
#> 913          Eulor1           DNA       DNA
#> 914          Eulor4        Eulor4   Unknown
#> 915          Eulor3           DNA      DNA?
#> 916         Eulor2C           DNA      DNA?
#> 917          LTR72B          ERV1       LTR
#> 918       LTR46-int          ERV1       LTR
#> 919          UCON21           DNA      DNA?
#> 920          MER136           DNA       DNA
#> 921          UCON17        UCON17   Unknown
#> 922           UCON6         UCON6   Unknown
#> 923          UCON20        UCON20   Unknown
#> 924          Eulor8         TcMar       DNA
#> 925         UCON28c       UCON28c   Unknown
#> 926           UCON8         UCON8   Unknown
#> 927         Eulor6C           DNA      DNA?
#> 928           UCON1         UCON1   Unknown
#> 929       MLT1G-int     ERVL-MaLR       LTR
#> 930          UCON19        UCON19   Unknown
#> 931           SATR2     Satellite Satellite
#> 932           LTR22          ERVK       LTR
#> 933         MER57E3          ERV1       LTR
#> 934           LTR70          ERV1       LTR
#> 935        BSR/Beta     Satellite Satellite
#> 936           HSAT4         centr Satellite
#> 937         Eulor6E           DNA      DNA?
#> 938           LTR35          ERV1       LTR
#> 939         UCON28b       UCON28b   Unknown
#> 940         Eulor10           DNA      DNA?
#> 941         UCON12A       UCON12A   Unknown
#> 942     HERVK13-int          ERVK       LTR
#> 943           L1P4c            L1      LINE
#> 944      HERV1_LTRd          ERV1       LTR
#> 945    HERVK11D-int          ERVK       LTR
#> 946           LTR7A          ERV1       LTR
#> 947    HERVFH19-int          ERV1       LTR
#> 948            MSR1     Satellite Satellite
#> 949     MLT1E1A-int     ERVL-MaLR       LTR
#> 950       SUBTEL_sa     Satellite Satellite
#> 951            GSAT         centr Satellite
#> 952          MER132         TcMar       DNA
#> 953          MER123           DNA       DNA
#> 954         Eulor6A           DNA      DNA?
#> 955           L1P4b            L1      LINE
#> 956     HERVL32-int          ERVL       LTR
#> 957          UCON25        UCON25   Unknown
#> 958       MLT1E-int     ERVL-MaLR       LTR
#> 959          UCON15        UCON15   Unknown
#> 960          Eulor7           DNA      DNA?
#> 961     HERVKC4-int          ERVK       LTR
#> 962         Eulor6B           DNA      DNA?
#> 963          UCON12        UCON12   Unknown
#> 964       LTR43-int          ERV1       LTR
#> 965            LSAU     Satellite Satellite
#> 966        (CATTC)n     Satellite Satellite
#> 967          REP522          telo Satellite
#> 968     HERV1_I-int          ERV1       LTR
#> 969           HSATI     Satellite Satellite
#> 970           MER9B          ERVK       LTR
#> 971  Charlie1b_Mars   hAT-Charlie       DNA
#> 972      HERV30-int          ERV1       LTR
#> 973       CheshMITE   hAT-Charlie       DNA
#> 974            SST1         centr Satellite
#> 975   Tigger1a_Mars  TcMar-Tigger       DNA
#> 976      HERV1_LTRc          ERV1       LTR
#> 977   HERV-Fc1_LTR1          ERV1       LTR
#> 978       HAL1N1_MD            L1      LINE
#> 979      HERV1_LTRe          ERV1       LTR
#> 980      HERV15-int          ERV1       LTR
#> 981   HERV-Fc1_LTR3          ERV1       LTR
#> 982   HERV-Fc1_LTR2          ERV1       LTR
#> 983    HERV-Fc2-int          ERV1       LTR
#> 984    HERV-Fc1-int          ERV1       LTR
#> 985      HERV1_LTRb          ERV1       LTR
#> 986       Charlie12   hAT-Charlie       DNA
#> 987  Cheshire_Mars_   hAT-Charlie       DNA
#> 988           HSAT6     Satellite Satellite
#> 989          D20S16     Satellite Satellite
#> 990       MLT1M-int     ERVL-MaLR       LTR
#> 991             SAR     Satellite Satellite
#> 992    Tigger2a_Car  TcMar-Tigger       DNA

# select family
rmskFilter(df, family="hAT-Charlie")
#>           gene_id   family_id class_id
#> 1           MER5A hAT-Charlie      DNA
#> 2           MER1A hAT-Charlie      DNA
#> 3           MER5B hAT-Charlie      DNA
#> 4         MER106B hAT-Charlie      DNA
#> 5           MER20 hAT-Charlie      DNA
#> 6      Charlie26a hAT-Charlie      DNA
#> 7          MER58C hAT-Charlie      DNA
#> 8           MER30 hAT-Charlie      DNA
#> 9       Charlie7a hAT-Charlie      DNA
#> 10          MER1B hAT-Charlie      DNA
#> 11         MER58B hAT-Charlie      DNA
#> 12        MER103C hAT-Charlie      DNA
#> 13         MER112 hAT-Charlie      DNA
#> 14      Charlie1b hAT-Charlie      DNA
#> 15         MER117 hAT-Charlie      DNA
#> 16      Charlie2b hAT-Charlie      DNA
#> 17      Charlie2a hAT-Charlie      DNA
#> 18         MER5A1 hAT-Charlie      DNA
#> 19          MER33 hAT-Charlie      DNA
#> 20       Charlie7 hAT-Charlie      DNA
#> 21         MER58A hAT-Charlie      DNA
#> 22     Charlie17a hAT-Charlie      DNA
#> 23     Charlie21a hAT-Charlie      DNA
#> 24     Charlie15a hAT-Charlie      DNA
#> 25      Charlie4z hAT-Charlie      DNA
#> 26        MER113A hAT-Charlie      DNA
#> 27           MER3 hAT-Charlie      DNA
#> 28     Charlie18a hAT-Charlie      DNA
#> 29       Charlie8 hAT-Charlie      DNA
#> 30     Charlie22a hAT-Charlie      DNA
#> 31          MER5C hAT-Charlie      DNA
#> 32        MER102c hAT-Charlie      DNA
#> 33      Charlie24 hAT-Charlie      DNA
#> 34         MER113 hAT-Charlie      DNA
#> 35      Charlie4a hAT-Charlie      DNA
#> 36     Charlie13b hAT-Charlie      DNA
#> 37        MER102b hAT-Charlie      DNA
#> 38       Charlie5 hAT-Charlie      DNA
#> 39      Charlie10 hAT-Charlie      DNA
#> 40     Charlie16a hAT-Charlie      DNA
#> 41         MER119 hAT-Charlie      DNA
#> 42     Charlie10a hAT-Charlie      DNA
#> 43     Charlie23a hAT-Charlie      DNA
#> 44     Charlie19a hAT-Charlie      DNA
#> 45       Charlie9 hAT-Charlie      DNA
#> 46     Charlie20a hAT-Charlie      DNA
#> 47         MER20B hAT-Charlie      DNA
#> 48       Charlie3 hAT-Charlie      DNA
#> 49      Charlie1a hAT-Charlie      DNA
#> 50        MER106A hAT-Charlie      DNA
#> 51      Charlie25 hAT-Charlie      DNA
#> 52         MER5C1 hAT-Charlie      DNA
#> 53     Charlie13a hAT-Charlie      DNA
#> 54        MER102a hAT-Charlie      DNA
#> 55       Charlie1 hAT-Charlie      DNA
#> 56         MER58D hAT-Charlie      DNA
#> 57       Cheshire hAT-Charlie      DNA
#> 58       Charlie4 hAT-Charlie      DNA
#> 59        MER113B hAT-Charlie      DNA
#> 60     Charlie14a hAT-Charlie      DNA
#> 61         MER30B hAT-Charlie      DNA
#> 62       Charlie6 hAT-Charlie      DNA
#> 63     Charlie10b hAT-Charlie      DNA
#> 64      Charlie11 hAT-Charlie      DNA
#> 65 Charlie1b_Mars hAT-Charlie      DNA
#> 66      CheshMITE hAT-Charlie      DNA
#> 67      Charlie12 hAT-Charlie      DNA
#> 68 Cheshire_Mars_ hAT-Charlie      DNA

# select class
df_subset <- rmskFilter(df, class= c("LINE", "SINE"))